updates id to be int instead of long

+++ services/AuctionEntryService.cs
@@ -14,7 +14,7 @@ public class AuctionEntryService(IDbService dbService) : IAuctionEntryService
         return true;
     }
 
-    public async Task<int> UpdateSoldAuctionEntries(string previousAuctionId, List<long> soldAuctionEntries)
+    public async Task<int> UpdateSoldAuctionEntries(string previousAuctionId, List<int> soldAuctionEntries)
     {
         return await dbService.EditData("""
             UPDATE auction_entry SET sold = true WHERE auction_id = @PreviousAuctionId AND id IN (@SoldAuctionEntries)
